Refactored PMTCT RRI linelists to a single report
